Move OidcLogo and OidcType into general components and adjust width for user access grid
