+++ user-interface/src/main/java/life/qbic/datamanager/views/login/LoginLayout.java
@@ -27,11 +27,11 @@ import java.util.List;
 import java.util.Map;
 import life.qbic.datamanager.views.AppRoutes;
 import life.qbic.datamanager.views.AppRoutes.Projects;
+import life.qbic.datamanager.views.general.oidc.OidcLogo;
+import life.qbic.datamanager.views.general.oidc.OidcType;
 import life.qbic.datamanager.views.landing.LandingPageLayout;
 import life.qbic.datamanager.views.notifications.ErrorMessage;
 import life.qbic.datamanager.views.notifications.InformationMessage;
-import life.qbic.datamanager.views.projects.project.access.OidcLogo;
-import life.qbic.datamanager.views.projects.project.access.OidcType;
 import life.qbic.datamanager.views.register.UserRegistrationMain;
 import life.qbic.identity.application.user.IdentityService;
 import life.qbic.identity.application.user.UserNotFoundException;
